Moving TestSettings from TestRun.cs to TestSettings.cs and ResultSummary from TestRun.cs to TestSettings.cs
