+++ 4b/Interfaces/src/TRX/TestDefinitions.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Xml.Serialization;
